requirements.txt
django==2.1.*
Pillow
djangorestframework
markdown
django-filter
requests